+++ cmd/gardener-extension-provider-local/app/options.go
@@ -39,7 +39,7 @@ import (
 	workercontroller "github.com/gardener/gardener/pkg/provider-local/controller/worker"
 	controlplanewebhook "github.com/gardener/gardener/pkg/provider-local/webhook/controlplane"
 	controlplaneexposurewebhook "github.com/gardener/gardener/pkg/provider-local/webhook/controlplaneexposure"
-	machinepodwebhook "github.com/gardener/gardener/pkg/provider-local/webhook/machinepod"
+	dnsconfigwebhook "github.com/gardener/gardener/pkg/provider-local/webhook/dnsconfig"
 	nodewebhook "github.com/gardener/gardener/pkg/provider-local/webhook/node"
 	shootwebhook "github.com/gardener/gardener/pkg/provider-local/webhook/shoot"
 )
@@ -68,7 +68,7 @@ func WebhookSwitchOptions() *webhookcmd.SwitchOptions {
 		webhookcmd.Switch(extensioncontrolplanewebhook.ExposureWebhookName, controlplaneexposurewebhook.AddToManager),
 		webhookcmd.Switch(extensioncontrolplanewebhook.WebhookName, controlplanewebhook.AddToManager),
 		webhookcmd.Switch(extensionshootwebhook.WebhookName, shootwebhook.AddToManager),
-		webhookcmd.Switch(machinepodwebhook.WebhookName, machinepodwebhook.AddToManager),
+		webhookcmd.Switch(dnsconfigwebhook.WebhookName, dnsconfigwebhook.AddToManager),
 		webhookcmd.Switch(nodewebhook.WebhookName, nodewebhook.AddToManager),
 	)
 }
